+++ src/main/groovy/dridco/seleniumhtmltojava/SeleniumBuilder.groovy
@@ -1,0 +1,8 @@
+package dridco.seleniumhtmltojava
+
+interface SeleniumBuilder {
+
+	def build(base)
+
+	def start()
+}